+++ src/qcommon/qcommon.h
@@ -729,7 +729,7 @@ void Com_Quit( void );
 int Com_ServerState( void );  // this should have just been a cvar...
 void Com_SetServerState( int state );
 
-unsigned Com_BlockChecksum( void *buffer, int length );
+uint32_t Com_BlockChecksum( void *buffer, size_t length );
 byte COM_BlockSequenceCRCByte( byte *base, int length, int sequence );
 
 float frand( void );  // 0 ti 1
